DeepLearning/NeuralStyleTransfer/requirements.txt
tensorflow-cpu
gradio
keras
matplotlib
numpy